breaking out generateCmds to own file
